BUG: Found two bugs in SETBC.f on Robin BC and cplBC
Major: On the Robin BC, when applied along the normal, the element normal
normal is not normalized. As a result, the applied traction is multiplied
by the area of the element and could lead erroneous results such as not
being treated at all as the values could be very low.

Minor: On the cplBC, while computing tangents numerically, only the flowrate
and pressures on the face that is perturbed are restored to original values.
However, when a flowrate on one face is perturbed, it could also change the
the flowrates/pressures on the other faces and these are not reset to
original values.