fixed velocity flux calc for struct/ustruct
Created a new function based on GNNB called GNNBT (T for current/deformed configuration). This function returns a normal vector weighted by the Jacobian, where the normal vector is the surface normal in the current configuration, and the Jacobian is the Jacobian of the mapping from the parent surface element to the current configuration surface element. Using GNNBT instead of GNNB in IntegV means integrating over the current configuration surface, instead of over the reference configuration surface. In particular, this fixes the velocity flux calculation when using struct (I think ustruct too). Before, the velocity flux was calculated as an integral over the reference configuration surface, which is not accurate (at least when using the velocity flux as the rate of change of enclosed volume)