added update tangent if flP, still working on it
I added a piece of code to BNEUFOLWP() (for follower pressure load) taken from FSILSINI() which updates the variable lhs%val to take into account the deformed geometry. I believe this val variable is the integral of the shape function times the normal vector over a surface, which is exactly what is needed to compute the cplBC tangent contribution (Moghadam et al. eq. 27). I added this piece of code so that these integrals are recomputed every Newton iteration.