Adding virtual face attempt
Adding code written to attempt to implement support for a virtual face. This code does not work. I ran into issues with the virtual face being partitioned incorrectly and information being inaccessible when running in parallel. After talking with Vijay, I will instead try to import the face as a member of the boundary condition structure.