adding support for virtual face
Adding code to compute integrals over virtual face. This requires communication between processors when integrals is performed. I have set it up so that master receives all the nodal info it needs to perform the integral, and performs the integral by itself. This could be change so that processors share the work of integration.