virtual face bug fixes and tangent matrix notes
In SETBC.f, added a line so that if a face is virtual and there is a Neumann or traction BC defined on it, it will cycle. Otherwise, this leads to a crash.

Also, when testing an idealized LV coupled to a constant flow rate LPN using genBC, I noticed the convergence was suboptimal (less than linear, when it should be quadratic). I think this is because the tangent matrix is inconsistent with the residual vector when a virtual face is used to compute flow rate. In this case, the integrals in the coupled BC tangent matrix contribution (see Moghadam 2013 eq. 27) are not the same. The right one should be taken over the capped surface (endo + virtual cap), while the other should be taken over the uncapped surface). I think this can implemented fairly easily in ADDBCMUL() by cycling in the second loop when the face is virtual. I just need to add and share the virtual flag to lhs%face. Currently, this doesn't compile, but this is what I will fix next.