virtual face tangent mat and debugging, not working
I believe when a virtual face is used to compute flow rate, the virtual face should have some contribution to the tangent matrix, even if not pressure is applied to it. Because no pressure is applied to it, the tangent matrix contribution should involve an integral over the capped surface and an integral over the uncapped surface. In ADDBCMUL(), I believe this amounts to ignoring the second do loop when the face is virtual. I attempted to do this, but I think the integral (in ValM), as well as the global indices in glob() may be incorrect when a virtual face is used when running in parallel, just like the previous issues I've had to sort out.

Before, the virtual face made no contribution to the tangent matrix because in genBC, I forgot to provide an offset value for the second coupled face (the virtual cap). This meant that the pressure on this face was always zero, so the resistance value for this face was always zero. In FSILS_SOLVE(), if res = 0 for a face, then that face is not considered coupled, so it doesn't contribute anything to the tangent matrix.