virtual cap face tangent contribution attempt
My latest attempt at improving the convergence of a struct-LPN simulation when a virtual face is used to cap. The basic idea was to modify the operation in ADDBCMUL() to take into account the virtual face. I had to make additions in other files to do this.

This does not work and I think I made an error. I can't figure out how to get the linear convergence that I saw before. I'm committing changes here and checking previous commits to see if I can get the linear convergence.