Fixing huge bug in ADDBCMUL()!!
If virtual face, want to skip its contribution to ADDBCMUL(), because it's contribution will be handled separately by the face that the virtual face caps. Before, used CONTINUE to skip in the DO loop if virtual face. CONTINUE is wrong. In Fortran, CYCLE is used to skip in a DO loop.

With this, I think the tangent matrix is appropriately constructed when a virtual face is used to cap another face. Before, I think the virtual face was contributing on its own to the tangent matrix, which ruined the tangent matrix and led to very poor convergence. With this change, convergence is excellent (quadratic or better).