adding catch in ADDBCMUL if cap has zero resistance
If cap has zero resistance, then the coupledFlag is FALSE, which means lhs%face(faIn) is not properly processed. In particular, the preconditioner routine does not run, which means valM is not filled in. If valM is accessed, this leads to undefined behavior and pollutes the tangent matrix, leading to no convergence.