bug fixes and create cap bc attempt
fixing a bug where the index in lhs%face(:) array for a cap surface is not computed correctly. Also, adding an attempt to automatically create a BC for a cap if a cap is specified in the input file under another coupled BC. I was able to make changes so that the cap bc is created, but as of this commit I still need to define a second coupled surface in genBC USER.f. I want the cap BC to take all information from its associated real coupled BC, but I can't figure out how to modify svFSI to do that properly